+++ pkg/config/load.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 Talend
+// Copyright © 2019-2020 Talend - www.talend.com
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,16 +23,16 @@ import (
 )
 
 // Load : Load Vault Sidecar Injector's config
-func Load(whSvrParams WhSvrParameters) (*InjectionConfig, error) {
+func Load(whSvrParams WhSvrParameters) (*VSIConfig, error) {
 	klog.Infof("annotationKeyPrefix=%s", whSvrParams.AnnotationKeyPrefix)
 	klog.Infof("appLabelKey=%s", whSvrParams.AppLabelKey)
 	klog.Infof("appServiceLabelKey=%s", whSvrParams.AppServiceLabelKey)
 
-	// Load sidecar config
-	var sidecarConfig SidecarConfig
-	err := loadYaml(whSvrParams.SidecarCfgFile, &sidecarConfig)
+	// Load injection config
+	var injectionConfig InjectionConfig
+	err := loadYaml(whSvrParams.InjectionCfgFile, &injectionConfig)
 	if err != nil {
-		klog.Errorf("Failed to load sidecar configuration: %v", err)
+		klog.Errorf("Failed to load injection configuration: %v", err)
 		return nil, err
 	}
 
@@ -65,11 +65,11 @@ func Load(whSvrParams WhSvrParameters) (*InjectionConfig, error) {
 		return nil, err
 	}
 
-	return &InjectionConfig{
+	return &VSIConfig{
 		VaultInjectorAnnotationKeyPrefix: whSvrParams.AnnotationKeyPrefix,
 		ApplicationLabelKey:              whSvrParams.AppLabelKey,
 		ApplicationServiceLabelKey:       whSvrParams.AppServiceLabelKey,
-		SidecarConfig:                    &sidecarConfig,
+		InjectionConfig:                  &injectionConfig,
 		ProxyConfig:                      proxyConfig,
 		TemplateBlock:                    templateBlock,
 		TemplateDefaultTmpl:              templateDefaultTmpl,